requirements.txt
streamlit
python-dotenv
PyPDF2
python-docx
langchain==0.2.1
langchain-community
langchain-huggingface



